refactor: 외부 API 호출 코드 패키지 정리 (#18)
