--- src/main/java/com/chocobi/leafy/constants/PlaceConstants.java
@@ -1,24 +1,0 @@
-package com.chocobi.leafy.constants;
-
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-public class PlaceConstants {
-    public static final String RESPONSE_TYPE_JSON = "json";
-    public static final String MOBILE_OS = "ETC";
-    public static final String APP_NAME = "Leafy";
-    public static final String MAX_NUM_OF_ROWS = "1400";
-    public static final String DEFAULT_NUM_OF_ROWS = "200";
-    public static final String PAGE_NO = "1";
-    public static final String BLANK = "";
-
-    public static final String AREA_PATH = "/areaCode1";
-    public static final String ECO_PATH = "/areaBasedList1";
-    public static final String RURAL_PATH = "/convergence2019/getConver04";
-    public static final String THEME_PATH = "/meta/KTOtour";
-    public static final String FARM_LIST = "/frmhsTasteHosList";
-    public static final String FARM_DETAIL = "/frmhsTasteHosDtl";
-    public static final String GEOCODE_PATH = "/v2/local/search/address.json";
-
-    public static final double[] DEFAULT_COORDINATES = {0.0, 0.0};
-}
