+++ src/main/java/com/chocobi/leafy/place/common/dto/Header.java
@@ -1,0 +1,12 @@
+package com.chocobi.leafy.place.common.dto;
+
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import lombok.Data;
+
+@Data
+@XmlAccessorType(XmlAccessType.FIELD)
+public class Header {
+    private String resultCode;
+    private String resultMsg;
+}
