+++ src/main/java/com/chocobi/leafy/place/common/dto/Items.java
@@ -1,0 +1,10 @@
+package com.chocobi.leafy.place.common.dto;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class Items<T> {
+    private List<T> item;
+}
