+++ src/main/java/com/chocobi/leafy/place/common/dto/Response.java
@@ -1,0 +1,9 @@
+package com.chocobi.leafy.place.common.dto;
+
+import lombok.Data;
+
+@Data
+public class Response<T> {
+    private Header header;
+    private Body<T> body;
+}
