--- src/main/java/com/chocobi/leafy/place/dto/common/Items.java
@@ -1,10 +1,0 @@
-package com.chocobi.leafy.place.dto.common;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class Items<T> {
-    private List<T> item;
-}
