--- src/main/java/com/chocobi/leafy/place/dto/common/Response.java
@@ -1,9 +1,0 @@
-package com.chocobi.leafy.place.dto.common;
-
-import lombok.Data;
-
-@Data
-public class Response<T> {
-    private Header header;
-    private Body<T> body;
-}
