--- src/main/java/com/chocobi/leafy/place/dto/farm/FarmDetailApiResponse.java
@@ -1,15 +1,0 @@
-package com.chocobi.leafy.place.dto.farm;
-
-import com.chocobi.leafy.place.dto.common.Header;
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlRootElement;
-import lombok.Data;
-
-@Data
-@XmlRootElement(name = "response")
-@XmlAccessorType(XmlAccessType.FIELD)
-public class FarmDetailApiResponse {
-    private Header header;
-    private FarmDetailBody body;
-}
