--- src/main/java/com/chocobi/leafy/place/dto/farm/FarmDetailBody.java
@@ -1,13 +1,0 @@
-package com.chocobi.leafy.place.dto.farm;
-
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlElement;
-import lombok.Data;
-
-@Data
-@XmlAccessorType(XmlAccessType.FIELD)
-public class FarmDetailBody {
-    @XmlElement(name = "item")
-    private FarmDetailItem farmDetailItem;
-}
