--- src/main/java/com/chocobi/leafy/place/dto/farm/FarmListItems.java
@@ -1,18 +1,0 @@
-package com.chocobi.leafy.place.dto.farm;
-
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlElement;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@XmlAccessorType(XmlAccessType.FIELD)
-public class FarmListItems {
-    @XmlElement(name = "item")
-    private List<FarmListItem> item;
-    private int numOfRows;
-    private int pageNo;
-    private int totalCount;
-}
