--- src/main/java/com/chocobi/leafy/place/dto/map/Document.java
@@ -1,10 +1,0 @@
-package com.chocobi.leafy.place.dto.map;
-
-import lombok.Data;
-
-@Data
-public class Document {
-    private String address_name;    // 주소
-    private String x;               // 경도
-    private String y;               // 위도
-}
