--- src/main/java/com/chocobi/leafy/place/dto/map/GeocodeResponse.java
@@ -1,10 +1,0 @@
-package com.chocobi.leafy.place.dto.map;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class GeocodeResponse {
-    private List<Document> documents;
-}
