+++ src/main/java/com/chocobi/leafy/place/entity/PlaceSourceType.java
@@ -1,0 +1,10 @@
+package com.chocobi.leafy.place.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public enum PlaceSourceType {
+    API, USER;
+}
