--- src/main/java/com/chocobi/leafy/place/entity/Type.java
@@ -1,10 +1,0 @@
-package com.chocobi.leafy.place.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public enum Type {
-    API, USER;
-}
