+++ src/main/java/com/chocobi/leafy/place/fetcher/farm/dto/FarmDetailBody.java
@@ -1,0 +1,13 @@
+package com.chocobi.leafy.place.fetcher.farm.dto;
+
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlElement;
+import lombok.Data;
+
+@Data
+@XmlAccessorType(XmlAccessType.FIELD)
+public class FarmDetailBody {
+    @XmlElement(name = "item")
+    private FarmDetailItem farmDetailItem;
+}
