+++ src/main/java/com/chocobi/leafy/place/fetcher/farm/dto/FarmListApiResponse.java
@@ -1,0 +1,15 @@
+package com.chocobi.leafy.place.fetcher.farm.dto;
+
+import com.chocobi.leafy.place.common.dto.Header;
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlRootElement;
+import lombok.Data;
+
+@Data
+@XmlRootElement(name = "response")
+@XmlAccessorType(XmlAccessType.FIELD)
+public class FarmListApiResponse {
+    private Header header;
+    private FarmListBody body;
+}
