+++ src/main/java/com/chocobi/leafy/place/fetcher/farm/dto/FarmListItems.java
@@ -1,0 +1,18 @@
+package com.chocobi.leafy.place.fetcher.farm.dto;
+
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlElement;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@XmlAccessorType(XmlAccessType.FIELD)
+public class FarmListItems {
+    @XmlElement(name = "item")
+    private List<FarmListItem> item;
+    private int numOfRows;
+    private int pageNo;
+    private int totalCount;
+}
