+++ src/main/java/com/chocobi/leafy/place/fetcher/kakao/dto/GeocodeResponse.java
@@ -1,0 +1,10 @@
+package com.chocobi.leafy.place.fetcher.kakao.dto;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class GeocodeResponse {
+    private List<Document> documents;
+}
