+++ src/main/java/com/chocobi/leafy/place/fetcher/tour/dto/EcoItem.java
@@ -1,0 +1,12 @@
+package com.chocobi.leafy.place.fetcher.tour.dto;
+
+import lombok.Data;
+
+@Data
+public class EcoItem {
+    private String addr;            //주소
+    private String mainimage;       //썸네일
+    private String title;           //제목
+    private String summary;         //설명
+    private String tel;             //전화번호
+}
